Add 102-fibanacci.c

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,7 +7,7 @@
 
 int main(void)
 {
-	int i, j, k;
+	long int i, j, k;
 	int counter;
 
 	i = 1;
